Move presence loop into misc cog
