+++ radia/cogs/misc.py
@@ -1,7 +1,7 @@
 """Misc cog."""
 
 import logging
-from random import randint
+import random
 
 import discord
 from discord.ext import commands, tasks
@@ -15,6 +15,7 @@ class Misc(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
         self.kraken.start()
+        self.update_presence.start()
 
     @commands.command(aliases=['🏓'])
     async def ping(self, ctx):
@@ -26,7 +27,7 @@ class Misc(commands.Cog):
     async def pet(self, ctx, num: int = None):
         """Get a picture of a pet."""
         embed = utils.Embed(title="Pets!", description="Picture of pets")
-        embed.set_image(url=f"https://cdn.vlee.me.uk/TurnipBot/pets/{num if num != None else randint(0, 140)}.png")
+        embed.set_image(url=f"https://cdn.vlee.me.uk/TurnipBot/pets/{num if num != None else random.randint(0, 140)}.png")
         await ctx.send(embed=embed)
 
     @commands.command()
@@ -46,6 +47,45 @@ class Misc(commands.Cog):
                 title="Error: Permissions missing",
                 description=f"Please find an IPL Staff member!")
             )
+
+    @tasks.loop(minutes=1)
+    async def update_presence(self):
+        """Loop to update the bot presence by selecting one of the strings at random."""
+        await self.change_presence(activity=discord.Game(random.choice([
+            "!help to get started",
+            # Signup!
+            "Signup for Low Ink!",
+            "Signup for Swim or Sink!",
+            "Signup for Testing Grounds!",
+            "Signup for Unnamed Tournament!",
+            # funny
+            "Powered by High Ink!",
+            "Investing in buying LUTI.",
+            "Get your coffee grounds 45% off this weekend at Testing Grounds.",
+            "Sink or Swim or Swim or Sink",
+            "According to all known laws of aviation",
+            # Round 4
+            "Round 4, here we go again!",
+            "The real round 4 were the friends we made along the way.",
+            # uwu stuff
+            "Sprinkles!",
+            "Wawa!",
+            # Socials
+            "Twitter: @IPLSplatoon",
+            "Twitch: twitch.tv/IPLSplatoon",
+            "Battlefy: battlefy.com/inkling-performance-labs",
+            "Patreon: patreon.com/IPLSplatoon",
+            "Github: github.com/IPL-Splat",
+            "Youtube: youtube.com/channel/UCFRVQSUskcsB5NjjIZKkWTA",
+            "Facebook: facebook.com/IPLSplatoon",
+            # People-specific
+            "Icon by Ozei!",
+            "Ban Kraken Mare",
+            "I kid you not Hoeen, he turns himself into a pickle",
+            "Go to sleep Lepto",
+            "Skye passed out again",
+            "Helpdesk needs you .jpg",
+        ])))
 
     @tasks.loop(hours=24)
     async def kraken(self):
